better handling of the cookies with = problem
