+++ src/main/java/org/radarbase/management/security/JwtAuthenticationFilter.kt
@@ -9,6 +9,7 @@ import org.radarbase.auth.token.RadarToken
 import org.radarbase.management.domain.Role
 import org.radarbase.management.domain.User
 import org.radarbase.management.repository.UserRepository
+import org.radarbase.management.web.rest.util.HeaderUtil.parseCookies
 import org.slf4j.LoggerFactory
 import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpMethod
@@ -120,26 +121,6 @@ class JwtAuthenticationFilter @JvmOverloads constructor(
         }
 
         return authHeader
-    }
-
-    /**
-     * Custom cookie parser as the httprequest.cookies method cuts off '='.
-     */
-    fun parseCookies(cookieHeader: String?): List<Cookie> {
-        val result: List<Cookie> = listOf()
-        if (cookieHeader != null) {
-            val cookiesRaw = cookieHeader.split("; ".toRegex()).dropLastWhile { it.isEmpty() }
-                .toTypedArray()
-            return cookiesRaw.map{
-                val parts = it.split("=".toRegex(), limit = 2).toTypedArray()
-                var value = if (parts.size > 1) parts[1] else ""
-                if (value.length >= 2 && value.startsWith("\"") && value.endsWith("\"")) {
-                    value = value.substring(1, value.length - 1)
-                }
-                Cookie(name = parts[0], value = parts[1])
-            }.toList()
-        }
-        return result
     }
 
     @Throws(IOException::class)
